chore: Update copyright year to 2026 and move deprecated code to archive
