+++ README.md
@@ -660,4 +660,4 @@ Feel free to contribute, report issues, or suggest improvements.
 
 License: [MIT](LICENSE)
 
-(c) 2025 [Zeix AG](https://zeix.com)
+(c) 2024 – 2026 [Zeix AG](https://zeix.com)
