+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@zeix/cause-effect",
-	"version": "0.16.2",
+	"version": "0.17.0",
 	"author": "Esther Brunner",
 	"main": "index.js",
 	"module": "index.ts",
